+++ packages/daemon/src/daemon.js
@@ -10,7 +10,7 @@ import { makeMailboxMaker } from './mail.js';
 import { makeGuestMaker } from './guest.js';
 import { makeHostMaker } from './host.js';
 import { assertPetName } from './pet-name.js';
-import { makeContextMaker } from './terminator.js';
+import { makeContextMaker } from './context.js';
 import { parseFormulaIdentifier } from './formula-identifier.js';
 
 const { quote: q } = assert;
